mtd: spi-nor: Update erasesize if dual parallel configuration
commit  1859edf811e7960ddc73a08b2e8bac93be463a87 from
https://github.com/Xilinx/linux-xlnx.git

For UBIFS we are using 64KB erase size that becomes 128KB for
dual parallel configuration. With the introduction of SFDP
the erase size is not updating as expected for dual parallel
configuration. This patch updates the erase size in dual
parallel configuration.

Signed-off-by: Tejas Prajapati Rameshchandra <tejas.prajapati.rameshchandra@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>