MLK-18279: ASoC: fsl_dsp: get the information of reserved memory from dts
commit  63823ab3e674068bd927ad35dbdbdc41a99d0998 from
https://source.codeaurora.org/external/imx/linux-imx.git

The reserved memory for dsp is defined in dts file, however, the dsp
driver has also defined the address and size of this reserved memory,
which is repeated and inflexible.

So by cancelling the definition in dsp driver and use system API to
get the information of reserved memory from dts dynamically to fix
this problem.

Signed-off-by: Weiguang Kong <weiguang.kong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>