MLK-18319-2 usb: cdns3: gadget: delete the unless code
commit  7d1a94b690d7218b2726b1b7797d2ef3bd68d1af from
https://source.codeaurora.org/external/imx/linux-imx.git

Confirmed from CDNS, this code is from bare metal, and not needed
for Linux. Below are their response:

This is alternative deferred interrupt.
This is intended to be used in Bare Metal to not block system in
interrupt.
This was not fully checked in Linux, so we do not recommend to use it
there.

Reviewed-by: Li Jun <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>