dmaengine: xilinx: dpdma: Align the line size with 128bit
commit 02fc44a63426c424ce451ecca38df997a0cd0eec from
https://github.com/Xilinx/linux-xlnx.git

The line size in the descriptor should be 128bit aligned per spec. Some
combinations of resolutions and pixel formats have non 128bit aligned
line size, and this patch rounds up the line size to be 128bit boundary.
The rounded line size value will always be within the stride which is
256byte aligned.

This doesn't add the alignment for other dma modes, cyclic and sg,
because there's a potential risk to break working cases. It should be
added later with extra care and testing.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Tested-by: Tejas Upadhyay <tejas.upadhyay@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <Satish.nagireddy.nagireddy@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>