irqchip/gic-v3-its: Move minimum LPI requirements to individual busses
At the moment, the core ITS driver imposes the allocation to be
in chunks of 32. As we want to relax this on a per bus basis, let's
move the the the allocation constraints to each bus.

No functionnal change.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
(cherry picked from commit 147c8f376e5526ebfaf2827734916414db52a3c4)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>