fix: irq: ap810: update the start interrupt of ICU in the GICv3
On next SoC revision of AP810 (B0), IOMMU context bank interrupt will
be wired to GICv3 entries #96-159.

This patch change the start interrupt of the ICU in the GICv3 to be 160
instad of 96, as preperation for B0.

In addition added note for GICP node:
In Armada-8K-Plus, the CP110 send message to the GICv3,
(instead of AP806 there's GICP unit that recive interrupt
from ICU and send to GIC).

The node of the GICP in AP810 is place holder until we get Marc Z.
patches that allow GICv3 to revice MSI interrupts from irq domain

Change-Id: I1fbc6a3511aea1c144bd03b04046cfae00db303a
Signed-off-by: Hanna Hawa <hannah@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>