pci: cp110: configure Resquester-ID to Stream-ID mapping
When a PCI devices access the AP internal bus it is assigned a
Stream-ID (SID) that identifies it in the System MMU (MMU-500). This
SID is calculated based on the PCI Requester-ID, a 16 bit value
composed of bus[15:8] device[7:3] and function[2:0]. Since the PCIe
Stream-ID is limited to 8 bits, a translation must take place.

This patch configures the SID mapping register to perform the
following mapping:
SID[7]   = 1. fixed prefix bit
SID[6:5] = bus[1:0]
SID[4:3] = device[1:0]
SID[2:0] = function[2:0]

The prefix is added to avoid a SID = 0 for bus = device = func = 0.
SID 0 is used by several units that wish to bypass the IOMMU like
the ICU so we need to avoid that.

Change-Id: I4a7a386d75aababeb9989ce143ff097d6cf0a434
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Signed-off-by: Hanna Hawa <hannah@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>