net: lan78xx: Disable TCP Segmentation Offload (TSO)
commit d08e75e53e7e501b0670b77ed1678ea43283b9a7 from
https://github.com/raspberrypi/linux.git

TSO seems to be having issues when packets are dropped and the
remote end uses Selective Acknowledge (SACK) to denote that
data is missing. The missing data is never resent, so the
connection eventually stalls.

There is a module parameter of enable_tso added to allow
further debugging without forcing a rebuild of the kernel.

https://github.com/raspberrypi/linux/issues/2449
https://github.com/raspberrypi/linux/issues/2482

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>