of: configfs: Use of_overlay_fdt_apply API call
commit 0d8b77f57811ff34a973b5d88d954eb29dad5cbd from
https://github.com/raspberrypi/linux.git

The published API to the dynamic overlay application mechanism now
takes a Flattened Device Tree blob as input so that it can manage the
lifetime of the unflattened tree. Conveniently, the new API call -
of_overlay_fdt_apply - is virtually a drop-in replacement for
create_overlay, which can now be deleted.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>