spi: spi-zynqmp-gqspi: Removed the interrupts not required to be handled.
commit a42ad124514b04710a0c2679233b5af1d5a535a1 from
https://github.com/Xilinx/linux-xlnx.git

For receiving data from the flash device we are enabling RXEMPTY and
RXNEMPTY interrupts out of this only RXNEMPTY interrupt we need to
check along with GENFIFIEMPTY interrupt. As we have enabled the
RXEMPTY interrupt but never handled there could be a situation
where isr will return IRQ_NONE and this will lead to spurious
interrupt. This patch fix this issue by not enabling the
RXEMPTY and TXEMPTY interrupts.

Signed-off-by: Tejas Prajapati Rameshchandra <tejas.prajapati.rameshchandra@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>