MLK-18626-2 ARM: dts: imx7d-sdb: Remove duplicate regulator-can2-3v3
commit  039dcc4c4043e9f6f355970c439ddcf4d10b3ee8 from
https://source.codeaurora.org/external/imx/linux-imx.git

Two different regulators are defined with the same name and label but
distinct properties.

The first definition was added with the first board dts and the second
was added when upstream added flexcan support.

Looking at schematics it is indeed gpio2 14 connected to the STB pin of
the CAN transceiver so remove the first definition. This also makes it
consisent with imx_4.9.y.

The second definition entirely overrides the first so this already
worked and this patch results in no DTB change, just a cleanup.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Fabio Estevam <fabio.estevam@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>