drm: xlnx: sdi: Fix the Vertical Field 1 Size calculation
commit  d5a752352b285bd5a87a77924dd483e20e8b1658 from
https://github.com/Xilinx/linux-xlnx.git

Fixes the Vertical Field 1 Size for Vertical Frame Size register in VTC.
For interlaced modes this is incremented by 1. But for 3GB modes, it is
incremented by 2. The 3GB mode is identified based on the DRM interlaced
flag and the pixel clock of 148.5MHz.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>