drm: xlnx: zynqmp_dpsub: Support reserved memory
commit  2cec4ca3a4868f65f54dc391e2d4dc4f980d59cc from
https://github.com/Xilinx/linux-xlnx.git

This adds support to use the assigned reserved memory for the buffer
allocation. The phandle to the reserved memory needs to be added to
the driver node using 'memory-region' property, as shown
in the example below.

	reserved: me@0x40000000 {
		compatible = "shared-dma-pool";
		reg = <0x0 0x40000000 0x0 0x10000000>;
		reusable;
	};

	zynqmp_dpsub: zynqmp-display@fd4a0000 {
		...
		memory-region = <&reserved>;
		...
	};

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>