drm: xlnx: zynqmp_disp: Modul param for initial format of gfx layer
commit  c7b46c6c691181cfbf8e6ee87c0550f5d697043d from
https://github.com/Xilinx/linux-xlnx.git

This allows the initial format for the graphics layer to be specified
and changed through module param. The format is used for
the fbdev emulation, and the emulated fbdev will be initialized with
the specified format: 0 = rgb565, 1 = rgb888, 2 = argb8888.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>