spi: spi-zynq-qspi: Added support to use gpio cs
commit a37f2445d1c174cbffa5689294e6fdf7eeae17e7 from
https://github.com/Xilinx/linux-xlnx.git

The spi-framework's cs-gpios property can support 3+ chip selects.
For gpio chip selects to work with the zynq-qspi controller,
we must notify the controller that a chip select event has occurred.
This can be done by poking the controller's first chip select
whenever a gpio chip select changes. But because of that, it isn't
possible to use gpio chip selects and the controller's chip selects
simultaneously. First controller chip select must be disconnected
as the controller will toggle the signal whenever a gpio chip
select changes. This patch adds support for that.

Signed-off-by: Brandon Maier <brandon.maier@rockwellcollins.com>
Signed-off-by: Clayton Shotwell <clayton.shotwell@rockwellcollins.com>
Signed-off-by: Tejas Prajapati Rameshchandra <tejas.prajapati.rameshchandra@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>