MLK-18767: drm: imx: dcss: fix vblank timed out warning
commit  437d6da13e64f2f380dad3ddf244e1f9b0ec5fce from
https://source.codeaurora.org/external/imx/linux-imx.git

The following commit:

f9a03b3 - MLK-17925: drm: imx: dcss: fix tearing

introduced a regression because we does not send VBLANK event if context
loader was not flushed. But this condition is always met immediately
after enabling the CRTC. Hence, when 30FPS modes are selected, it is
possible to exceed the 50ms vblank timeout if the first event is not
sent. That's because 2 x 0.33ms > 50ms. This does not happen for 60fps
modes though.

This patch will force sending the first vblank event even if the context
loader is not completely flushed.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>