MLK-18877-3: arm64: dts: imx8mq: Refactor dts files
commit  9335aa842c5bef1230223461e6413ffeabbc5321 from
https://source.codeaurora.org/external/imx/linux-imx.git

Currently, there are a lot of dts bindings duplicated into many files
because of the multiple possible use-cases.
Still, some bindings can be defined in a single file, then re-used in
each specific dts file, with the required specific scenario changes.
Therefore:
1. Move the adv7535 (DSI-HDMI converter) and synaptics_dsx_ts (DSI touch)
definitions to fsl-imx8mq-evk.dts file.
2. Removed the max-res binding from the main dtsi file. This is a
limitation and should NOT be present in the main definition of lcdif.
Use this limitation only in specific dts files which requires such
limitations. For example in the dual-display.dts file, where the lcdif
is used along with the dcss; lcdif has to be limited because of the low
bandwidth when used in the same time with dcss.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>