soc: ti: add HWMOD_INIT_NO_IDLE to workaround the hardware issue
On am3358, there is a bug in silicon that clocks cannot be cut to DEBUGSS.
It has been confirmed from the following vendor website:

https://e2e.ti.com/support/arm/sitara_arm/f/791/p/713673/2632160#2632160

A kernel fragment AM335X_DEBUGSS_NO_IDLE is introduced to workaround it
in case this hardware issue might be OK in the latest version. Currently
it is enabled by default in scc/cfg.

Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
(cherry picked from commit a8cf9b9cdd1267fefe9332224278f622032f14eb)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>