MLK-19151: ASoC: fsl_asrc: fix dma timeout issue for imx8qxp
commit  3519b67ac01ccd7fb0fd251273c7017f1404c869 from
https://source.codeaurora.org/external/imx/linux-imx.git

In the imx8qxp, the DMA is EDMA, which require the buffer size
should be divided by burst size with no remainder, otherwise
the remainder is not transferred by EDMA, so the input buffer
is not consumed by ASRC, then there will be dma output timeout
issue. This behavior is different with SDMA.

This patch is to change the input burst size to be 1 to avoid
the issue, and refine the last_period_size for output buffer.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>