net: octeontx2: Attach RVU NIX and NPA block LFs
For a RVU PF to function as a NIC, NPA (for receive buffers)
and NIX (for rcv, send and completion queues) are the
minimum resources needed. So request admin function (AF)
to attach one each of NIX and NPA LFs (local functions).
Also get the msix vector offsets for these LFs.

Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>