net: mvpp2: fix 10G-SFP link status interrupt allocation
PROBLEM:
On 8040-DB/7040-DB boards the 10G SFP managed "link" status interrupt
is present but doesn't work. So 10G devices are not Up, or randomly Up
and do not detect the fiber-cable connect/disconnect.

Root-cause:
Wrong condition deciding to Not allocate link-status interrupt.

Solution:
Go to allocate the interrupt if phylink is present but the "has_phy"
is FALSE. This is exactly the case of the SFP.

Change-Id: Ie1dacafd20a1ef48f0b8ccb32ecad2a56e236bad
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/59471
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: Stefan Chulski <stefanc@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>