MGS-4111 [#imx-1117] Fix displaying wrong governor values on 8QP
commit  e412173eb3ed4b9383ba324ad1837a40ad06a8c2 from
https://source.codeaurora.org/external/imx/linux-imx.git

On 8QP we don't have the OVERDRIVE governor, only NOMINAL and
UNDERDRIVE.  We set the current initial governor to OVERDRIVE, but also
display incorrect governor names relative to the clocks assigned in OPP
tuples.  This patch address this situation by determining which governor
should be used at start-up and fixes displaying the governors
available for use.

This patch also sets the initial clock rate, in case the assigned-clock
rates for the cores are different than the ones found in the
operating-points (OPP). While this shouldn't be a real use-case this
allows testing different OPP tuples to mimic a OPP array for 8QP.

Signed-off-by: Marius Vlad <marius-cristian.vlad@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>