net: xilinx: axiethernet: Fix memory leak in axienet bd_free()
commit a0d8d595d70da4f566d346b594c619a01e143c13 from
https://github.com/Xilinx/linux-xlnx.git

There is a memory leak in DMA handling functions that lead to kernel
crash after several cycles of interface up-down operations. This leak
was introduced in 'commit 99d053bb0e87 ("net: ethernet: Fix issues in
the driver when DRE is not enabled in the h/w")' by allocating a pool
of TX buffers.

This patch add forgotten dma_free_coherent call on resources deallocation.

Signed-off-by: Andrey Kartashev <a.s.kartashev@gmail.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>