net: mvpp2: allow to assign a reserved memory region for DMA allocations
This patch adds the possibility to reserve a DMA region in the device
tree, and to assign this memory region to the PPv2 device for DMA
allocations. This helps to avoid cases where the default CMA pool is to
small in the kernel, by having a PPv2 specific pool we can control.

Change-Id: I3326471fbda897dcfd22a5b06034b0723d2bd457
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/59465
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: Stefan Chulski <stefanc@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>