net: mvpp2: use tx bulk writing to aggregated tx-queue
The traffic performance could be improved (up to 22%) by reducing
Aggregated TX-queue counter register (write a bulk of N instead of
one-by-one). This requires defer in xmit-procedure and bulk-timer.

This patch modifies the xmit-procedure to defer, accumulate N tx-packets
and write to Aggregated queue this bulk.
The 50uSec bulk timer is used to finalize the xmit in case of traffic
stop or gap.
The feature is enabled by default.

Change-Id: I3478b5c9ce0013279ccf0d86f38f8a625d74a2a4
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/59486
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: Stefan Chulski <stefanc@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>