enetc: fix writeback bit of TX BD
The W bit of TX buffer descriptor is used to determine
if writeback occurred after completion of this frame.
It will be checked before reading TX timestamping.
This patch is to fix the definition of this bit becasue
the value was wrong and the name was not very proper.

Signed-off-by: Yangbo Lu <yangbo.lu@nxp.com>
(cherry picked from commit 44643ae18bc59fc7c41d5f69524ad5f175ac2d16)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>