soc: octeontx2: Map or unmap NPC MCAM entry and a counter
Alloc memory to save MCAM 'entry to counter' mapping and since
multiple entries can map to same counter, alloc memory for
tracking counter's reference count as well.

Do 'entry to counter' mapping when a entry is being installed
and mbox msg sender requested to configure a counter as well.
Mapping is removed when a entry or counter is being freed or
a explicit mbox msg is received to unmap them.

Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>