soc: octeontx2: Alloc and config NPC MCAM entry at a time
A new mailbox message is added to support allocating a MCAM entry
along with a counter and configuring it at a time. This reduces
the amount of mailbox communication involved in installing a new
MCAM rule.

Also added a msg to retrieve NPC Rx and Tx interface's KEY
configuration i.e key width and KEX nibble config. This info
is used by PF/VF drivers to frame MCAM entry data.

Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>