soc: octeontx2: Add resource partitioning config
This commit adds facilities for configuring and tracking resource
partitioning between RVU PF devices (and indirectly their related
RVU VFs).
The AF keeps track of the PF level resource partitioning only because
partitioning between VFs is a job of each RVU PF and only RVU PF driver
knows the moment and number of VFs enabled.

For each tracked resource (currently RVU LFs), user can set a limit
for each RVU PF on how many resources may be attached to it via
MBOX (MBOX_MSG_ATTACH_RESOURCES msg for LFs).
This limit value K for PF X means that the sum of attach requests
for that resource coming from all pcifunc belonging to PF x (called
PF x's family) cannot exceed K.
The sum of limit values for a particular resource always has to be lower
or equal to the total amount of those resources on the system.

The limits are modifiable for PFx until resource attach message is
received. At that point they become read-only until all resources are
detached from that device.

The default values for the limits are defined as follows:
* RVU AF (for its future VF handling)
  - NIX: pci_sriov_get_totalvfs(pdev)
  - NPA: pci_sriov_get_totalvfs(pdev)
* networking RVU PF
  - NIX: 1 + pci_sriov_get_totalvfs(pdev)
  - NPA: 1 + pci_sriov_get_totalvfs(pdev)
  - rest: 0
* SSO/TIM RVU PF
  - NPA: pci_sriov_get_totalvfs(pdev)
  - SSO: N(SSO_LFS) / N(SSO_PFS)
  - SSOW: N(SSOW_LFS) / N(SSO_PFS)
  - TIM: N(TIM_LFS) / N(SSO_PFS)
  - CPT: num_online_cpus() / N(SSO_PFS)
  - rest: 0
* NPA RVU PF
  - NPA: pci_sriov_get_totalvfs(pdev)
  - rest: 0
* CPT RVU PF
  - CPT: num_online_cpus()
  - NPA: 1
  - rest: 0

Signed-off-by: Stanislaw Kardach <skardach@marvell.com>
Signed-off-by: Krzysztof Garczynski <kgarczynski@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>