MLK-19413-19 gpu: imx: dpu: framegen: Do not warn on zero pclk rate for slave framegen
commit  379f5b1439fb4a1ae46c0232d6da058c6c267b55 from
https://source.codeaurora.org/external/imx/linux-imx.git

The pixel clock of the slave framegen comes from the master framegen
when the two framegens work in sync mode, so we cannot get pixel clock
from the pixel clock rate from the slave framegen directly to calculate
the timeout value when waiting for it's done.  Instead, we can tell
the rate from the video mode.  So, we take this fallback way to get
the rate in this case and don't warn.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>