drm/bridge: hw-hdmi: Control clock of bridge during S2RAM
commit 02b6acf4089e00b0e90e268192147255e6bafca9 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

This patch helps to control the clocks of
bridge/hw-hdmi during system suspend/resume.

Currently, "hdmi" clock is enabled when the DU driver
is enabled but it is not disabled at system suspend.

So, when the system resumes the software state (enabled)
and hardware (disabled) of clock are not matched.
And it makes the system resume fails.

This patches is necessary to synchronize software
and hardware state of clock when backup/restore code
of clock (CPG module) is removed.

In addtion, the procedure for enabling hot plug is added.

Reported-by: Dien Pham <dien.pham.ry@renesas.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>