rpi-wm8804-soundcard: use nicer driver_name "RPi-WM8804"
commit 8d77fa9024e4f7d108a70f8bb5384eaadb9561f8 from
https://github.com/raspberrypi/linux.git

The driver_name shows up in /proc/asound/devices and is used
to load the proper Alsa card conf from /usr/share/alsa/cards.

But the length is limited to 15 characters and spaces are
replaced by underscore so the mangled name currently is
"RPI_WM8804_soun" which is not very nice.

Use "RPi-WM8804" instead which won't be mangled and can be used
unaltered as the card conf filename and card name in that file.

Signed-off-by: Matthias Reichl <hias@horus.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>