MLK-19435 soc: imx8: pm-domains: Report failures to pm core
commit  3d81ae2bb96a5c3c74a90982f872885d05f13b2c from
https://source.codeaurora.org/external/imx/linux-imx.git

Right now power_on always returns true even if SCFW reports a failure.
Since the target resource is still unpowered this quickly turns into a
hang when we attempt to access it.

Handle this by reporting an error to the PM core instead and also print
the sc_err number to help with debugging.

This fixes boot on 8qm A0: instead of hanging on boot we print an error
and refuse to probe the VPU.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Dong Aisheng <Aisheng.dong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>