drm: rcar-du: Fix possible crtcs for plane when using VSPDL
commit 653e1374d0389bd4f1bd6f4005a3770b802d37c8 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

If two or more channels are specified for possible crtc of plane,
if user move the overlay between monitors in the system test,
the following error occurs.

switching CRTC directly
atomic core check failed

In order to solve the problem, this patch fixes it by fixing
assignment of possible crtc for each plane.

In addtion, add an option to change the assignment of the
overlay sharing with VSPDL.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>