media: rcar-vin: Add streaming count check when streming off
commit b762b33eeab851d5c4d43f95336c689769f56aa7 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

When executing suspend to ram, in the case of abnormal termination,
media_pipeline_start is not executed and media_pipeline_stop only
may be executed in some cases.
In order to prevent it, this patch adds a check of stream count.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>