drm: bridge: thc63: Restrict modes based on hardware operating frequency
commit ce051c65f62baf9a3926449589686c272fb9777a from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

The THC63LVD1024 is restricted to a pixel clock frequency in the range
of 8 to 135 MHz. Implement the bridge .mode_valid() operation
accordingly.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>
Link: https://patchwork.kernel.org/patch/10587279/

In addition, Refresh rate 30 or less of full HD is basically not used,
so remove it.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>