dma: axi-dmac: don't check the number of frames for alignment
It doesn't look like the 2D transfers were used much, prior to commit
81b810593ebcf1a537042704c8363bbaafbcba0b, so this may have not received
much attention.

There is no requirement for Y_LENGTH to be aligned to the bus-width (or
anything). X_LENGTH is required to be aligned though.

So, we shouldn't check that the number of frames is aligned.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 04231a9a3ff27630f3453b5feca197723a5f3e88)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>