PCI: aardvark: Add support for comphy initialization
On Armada 37xx platforms the comphy (common phy) must be configured to
work in appropriate mode. Hitherto comphy configuration for pcie was
done by bootloader. As a comphy driver was extended to support pcie
comphy configuration, this patch uses it in pcie driver. Thanks to that
there is no need to rely on the bootloader configuration which could
configured the comphy to different mode.

Change-Id: I3c51d30e90c91debfa0d87f0ed015057e783f878
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>