drm: rcar-du: Fix dpad signal routing for R8A7799X
commit 99547173de6e8e1ab5ef1505bf1ec0eb736c4bca from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

In the case of R8A7799X, the dpad0_source has an initial value of 0
when starting kernel, DPAD can not be set with DU1 when using with
LVDS0 -> HDMI (DU0). So correct it so that it searches and decides
connector information of VGA.

In addition, The default value of dpad0_source is determined to be 1,
and the default setting is also changed.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>