i2c: rcar: Fix I2C DMA transmission by setting sequence
commit 55d2d2fb8b04b4b7bbdbd68d5e0523841b610de2 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

By clearing the i2C MDE bit as the HW specification, it is necessary
to start transmission after 1 byte PIO transmission.

For 1st byte, it is necessary to perform PIO transfer.
Transfer is started by clearing the MDE bit, but I think that there
is a possibility that DMA transfer will be started before that.
In accordance with the HW specification, the transfer of 1st byte
was corrected so as to clear MDE clearly.
DMA transfer is used after the second byte.

Signed-off-by: Hiromitsu Yamasaki <hiromitsu.yamasaki.ym@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>