net: octeontx2: fix msix allocation/release
This fixes kernel oops in otx2_disable_msix().  For this function to work
correctly it should be called when nix_msixoff is set.  It is not true
in every error path.

Split the pair of functions otx2_enable_msix()/otx2_disable_msix() into two
parts:

1.
- Allocation of memory for irq_name and affinity_mask
- pci_alloc_irq_vectors()
- Initialization of MBOX interrupt

These can be done in the _probe() function and undone in _release() so for
memory allocation we can use the functions for managed allocation.
Also this means that MBOX interrupt is functional for the whole life time
of the driver.  That allows to get rid of otx2_hw->num_vec and therefore
of otx2_nic->set_mac_pending.

2.
- Request/free of CINTs.

This should be done in _open()/_stop().  Requesting was at _open()
initially so free_irq() was moved to otx2_free_cints()

Some small improvements have also been done:

- pci_enable_device() -> pcim_enable_device()
- sprintf() -> snprintf()
- shorten some expressions by using a local variable for interrupt name
  at some place

Signed-off-by: Aleksey Makarov <amakarov@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>