octeontx: mmc: fix setting clock for non-zero bus IDs
There is an issue in all Octeon and OcteonTX devices where the clock
settings in the MIO_EMM_SWITCH register only take effect if the BUS_ID
field is zero.

For any device other than device zero this means that the register must
be written twice, first with BUS_ID set to zero and again with the proper
BUS_ID in order for the CLK_LO and CLK_HI fields to take effect.

Signed-off-by: Aaron Williams <aaron.williams@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>