MLK-19616 sound: asoc: fix quality selection bug
commit  3afb3d9fc6a7588787e002fad9a865607f0c27b1 from
https://source.codeaurora.org/external/imx/linux-imx.git

The quality selection bitfield is represented as following:
 * Medium»      0 0 0
 * High»»       0 0 1
 * Very Low 2»  1 0 0
 * Very Low 1»  1 0 1
 * Very Low 0»  1 1 0
 * Low» »       1 1 1

One might notice that from High to Very Low 2 quality there is
a gap of 2 (010 and 011) are missing. In the previous implementation
of amixer control, when you are selecting other quality than
the default mode, the value will always jump from High to Low
with a 2 step and will not be able to go back to a previous value.
The solution was to add two "N/A" undefined values in the
selection texts since those two values are marked as reserved
in the micfil RM.
The alternative was to write custom get and put functions to
only do the translation but the aproach in this patch is easier
to understand and less error prone.

Signed-off-by: Cosmin-Gabriel Samoila <cosmin.samoila@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>
(cherry picked from commit 976f005b20ddc12c7fe0ce60edd99e5fb9400d21)
Signed-off-by: Meng Li <Meng.Li@windriver.com>