misc: xilinx-sdfec: Include PL init in irq flow
commit a42b2c654fa488da6e33d6df1764dcdf2cc62a6b from
https://github.com/Xilinx/linux-xlnx.git

Extend the interrupts monitored by the driver to include ECC PL
initialization interrupts. Update state to PL_RECONFIGURE if a
multi-bit error is detected in this range.
These interrupts are used only when the SD-FEC block is configured and
have coefficients initialized by the Programmable Logic. If an ECC PL
initialization multi-bit error is detected, the PL must be
re-configured to reload the coefficients.

Signed-off-by: Mauro DeLuca <mauro.de-luca@xilinx.com>
Signed-off-by: Derek Kiernan <derek.kiernan@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>