mtd: nand: pl35x: Add ECC checking for ondie nand flash
commit dce4781567cbd67abb87ef3fc708b7fdd4ce6838 from
https://github.com/Xilinx/linux-xlnx.git

Check status of the operation and increment the correctable or
uncorrectable error counters which is critical for UBI layer to greatly
improve handling of the errors.  In particular it used to trigger migrating
the data to safely off before scrubbing the failing Erase Blocks. Even
uncorrectable errors may often be handled by UBI if it is in a
non-critical page such as Erase Counter header.

Signed-off-by: Andrew Worsley <amworsley@gmail.com>
Reviewed-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>