vchiq_2835_arm: Implement a DMA pool for small bulk transfers (#2699)
commit 810cf924b67992f5d0b21e10458ad53062311638 from
https://github.com/raspberrypi/linux.git

During a bulk transfer we request a DMA allocation to hold the
scatter-gather list.  Most of the time, this allocation is small
(<< PAGE_SIZE), however it can be requested at a high enough frequency
to cause fragmentation and/or stress the CMA allocator (think time
spent in compaction here, or during allocations elsewhere).

Implement a pool to serve up small DMA allocations, falling back
to a coherent allocation if the request is greater than
VCHIQ_DMA_POOL_SIZE.

Signed-off-by: Oliver Gjoneski <ogjoneski@gmail.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>