dma: axi-dmac: populate residue info for completed xfers
This change implements computing DMA residue information for transfers.

The way this is done, is to dequeue the partial transfers from the FIFO of
partial transfers, store the partial length to the correct segment &
descriptor, and compute the residue before submitting the DMA cookie to the
DMA framework.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 31baa60b75629fe466ddf8c5cbd16b7156692358)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>