dma: xilinx: Mark previous active decriptor as complete in irq handler
commit 7ddb98fd2bea3fac72503a9c76c3fc9ca12957e5 from
https://github.com/Xilinx/linux-xlnx.git

Moving the block to mark previous active descriptor as complete
to irq handler as it's more appropriate to do it in interrupt
context which confirms that previous transaction was completed.

There is no semantic change in the flow,
previously descriptor was marked as completed in
xilinx_frmbuf_start_transfer() assuming it's is called by
the done interrupt handler which didn't seem required,
so we now mark it as completed in the interrupt handler itelf.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Reviewed-by: Satish Kumar Nagireddy <satish.nagireddy.nagireddy@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>