crypto: zynqmp-aes: Fix for segfault seen with large sets of data
commit 083facadcf906d31fd3308bd372a83d31534b49d from
https://github.com/Xilinx/linux-xlnx.git

This patch resolves the issue seen when encrypt/decrypt of large sets of
data.

Issue is seen due to improper handing of blkcipher_walk. When large
amount of data is sent for encrypt/decrypt, af_alg framework copies it
to the TSGL (TX SGL) kernel. When driver does a walk on the scatterlist,
walk returns nbytes to be processed. nbytes can be complete data sent
from user space or few bytes. So, when data is received in multiple
walks, we are accumulating the data and sending complete data at once.

Signed-off-by: Kalyani Akula <kalyani.akula@xilinx.com>
Reviewed-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>