media: rcar-vin: Fix clock control flow
commit c00dec285f5294eedf69f38eae7df5e0c92628db from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

About VIN drivers sequence at using CSI2, it is nesessary that
CPG (Clock Pulse Generator) reset/release is added according to
H/W manual. So this patch adds it.
In addition, CPG stop status check. The start check of CPG is
unnecessary because the CPG enable side is already supported
by the Clock driver.

Also, if the enable/disable of CPG is implemented with an open or
release function, it may be called wastefully from another process,
so this patch modifies it to control CPG when Stream ON/OFF.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>