MLK-19854-1: ASoC: imx-cs42888: fix error when m4 image is not loaded
commit  d4eb8ab2639947cd486457a878ce57d2d1dd2c51 from
https://source.codeaurora.org/external/imx/linux-imx.git

With rpmsg sound, when the m4 image is not loaded. There is error log

[ 46.257647] imx-cs42888 sound-rpmsg-cs42888: ASoC: CODEC DAI cs42888 not registered
[ 46.265413] imx-cs42888 sound-rpmsg-cs42888: snd_soc_register_card failed (-517)

The issue is that codec is not register, the sound registeration will fail

So add check in probe function for codec dai ready or not.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>