media: ov5640: set DPHY timing registers
commit 4077811faa00bb56f6147f1964ad90c01ca9efdd from
https://github.com/Xilinx/linux-xlnx.git

Register 0x4837 is used to calculate the DPHY timing
parameters such as HS_SETTLE. It needs to be set to the
current pclk period. It is possible for some CSI receivers
to tune this on the RX side, but some do not.

Currently this value is not set, so fix this by setting it
whenever setting the rest of the PLL settings.

Also, a change to the init register blob that helps ensure
the following DPHY spec requirement is met:

MIN HS_ZERO + MIN HS_PREPARE > 145 + t_UI * 10

Signed-off-by: Sam Bobrowicz <sam@elite-embedded.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>