mmc: renesas_sdhi: Change HW adjustment register according to speed mode
commit d286828a8ea35374297425c03f5ecd27fac35e10 from
git://git.kernel.org/pub/scm/linux/kernel/git/horms/renesas-bsp.git

SCC register is used for SDR104/HS200/HS400.
It is necessary to change the setting value of SCC_DT2FF
according to the speed mode.
If this value is inappropriate, CRC error tends to occur.

This adds variable "tap_hs400" for HS400 mode.
And controls register value according to speed mode.

This information will be reflected in the R-Car Gen3
Hardware Manual in the future.
Refer to it for details.

Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>