dma: xilinx: Add early callback support to Framebuffer Driver
commit 99df3ae2b6ba6d24c7a790768998118a582242ef from
https://github.com/Xilinx/linux-xlnx.git

This patch adds support for the callback per descriptor to be called
after the staging phase itself rather than waiting for the descriptor
to transition to the active stage and done list.
This behavior is enabled by an API which lets dma engine client decide
whether the callback needs to called as usual or earlier after staging
state.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>