MLK-19936-2 arch: arm64: fsl-imx8mm: add audio plls in micfil node
commit  66efd66691cb3be3580199e85d704cb996ce2f0b from
https://source.codeaurora.org/external/imx/linux-imx.git

Parse the clock tree and stop when you find one of the
audio_plls. If this audio_pll cannot support the required
rate, change the parent to the other pll.
Set rate to rate * 1024 so we can support all parameter
configurations with a minimum clock rate.
This is required to support all rates multiple of 11025 and 8000.

Signed-off-by: Cosmin-Gabriel Samoila <cosmin.samoila@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>