octeontx: change count threshold for vhaura
When FPA_VHAURA_CNT is equal or crosses FPA_VHAURA_CNT_THRESHOLD then
A_THRESH interrupt is generated. Previously FPA_VHAURA_CNT_THRESHOLD
was set to half of the numbers of buffers in a pool. This patch
changes FPA_VHAURA_CNT_THRESHOLD to a number of buffers in a pool
minus one.

Signed-off-by: Lukasz Bartosik <lb@semihalf.com>
Signed-off-by: Yury Norov <ynorov@marvell.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>