octeontx: add and handle start domain message
The purpose of start domain message is to inform kernel driver
that a given domain is being used by an odp application and while
being used it can not be destroyed. Kernel driver clears the flag
preventing domain destruction on each domain reset. It also stores
jiffies of last domain reset and does not allow to destroy domain
within 1 second since last domain reset to account for time difference
when odp application opens devices and a time when it actually sends
start domain message to the kernel driver.

Signed-off-by: Lukasz Bartosik <lb@semihalf.com>
Signed-off-by: Yury Norov <ynorov@marvell.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>