net: mvpp2: fix: no bulk tx for local-host packets
Problem appearance:
iperf-UDP reports Out-Of-Ordering on the server side.

INFO:
For performance improving (up to 20%) the TX sends a bulk of packets
with defer/bulk timer.
Round-robin on hw packet sending from TX-queue + CPU-affinity-switch
(on iperf TX) could cause Out-Of-Ordering.
The RX-to-TX packets are not in this case since both RX and TX are
always done on same softirq context / same CPU.

FIX:
Check the RXTX hash (placed by RX) for the defer decision.
Do not bulk/defer the packet without this hash and flush the bulk at
once.

Change-Id: I22e889651cc3b03a126f6582035a3d682158e9f3
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/60525
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>