net: octeontx-fpa: fix sysfs VF indexing
pdev->devfn is offset from the pci_iov_virtfn_devfn(fpa->pdev, i) so
the indexing of VF device structures in _show() functions was offset
by 1 which caused invalid values to be shown.

Additionally there was a possibility of a NULL offset read because of
the indexing above.

This commit fixes the indexing offset to retrieve a proper device
structure. This applies for FPA device files: pool_maxcnt, pool_curcnt
and pool_redcnt.

Signed-off-by: Stanislaw Kardach <stanislaw.kardach@cavium.com>
Signed-off-by: Yury Norov <ynorov@marvell.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>