misc: otx_bphy_ctr: Add OcteonTx2 BPHY control driver
This patch adds OcteonTx2 BPHY control driver. This driver is used to
setup interrupt handlers for BPHY device in firmware through SMC calls.
The interrupt is handled by EL3 firmware and redirected to user space
handler that are registered.

Signed-off-by: Radha Mohan Chintakuntla <rchintakuntla@cavium.com>
[Kevin: The original patch got from Marvell sdk10.0_19.06.
Just some minor context mods in order to port to wrlinux]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>