MLK-19750-5: ARM64: dts: workaround 192kHz recording noise issue
commit  e818f109021f1257635393b5849b669bb47a52bf from
https://source.codeaurora.org/external/imx/linux-imx.git

There is noise issue with 192kHz recoding with ESAI + CS42xx8 on imx8qxp
mek board.

This issue is caused by the round trip delay due to longer trace length
on board. After we switch to tx master, rx slave mode, the issue is gone.
so the setting can workaround the issue, the reason is that the bitclock,
frame clock and data is generated from one side, for recording, is from
codec, the asynchronous of clock and data is eliminated.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>