staging: xroeframer: add support for multi-port Ethernet
commit b2e37c37245f8bfbe472c54e122834782c837250 from
https://github.com/Xilinx/linux-xlnx.git

Support for up to 4 Ethernet ports. This wraps all the "store" & "show"
functions of sysfs entries in a function which calculates the current
Ethernet port (when applicable) and adds the correct offset to the
address' register, which is ADDRESS + (0x100 * port_num). It defaults
to 0 for the non-applicable cases so no other entries are affected.

Signed-off-by: Vasileios Bimpikas <vasileios.bimpikas@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>