MLK-20242: drm/imx/dcss: fix brightness for REC.709
commit  1e3fcb39a6fca44bae78a12b50c7112dc95e96ae from
https://source.codeaurora.org/external/imx/linux-imx.git

The brightness, when REC709 was used in the configuration of the pipes,
was lower than expected.

The reason was the HDR10 configuration application that was used to
create the tables had a parameter that was wrongly set.

The tables were re-generated with the proper setting.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>